corrected the formula to compute the reduced mass
When using the FEYNMAN_HIBBS_LENNARD_JONES potential (and its other variations) to compute the van der Waals potential, the formula to compute the reduced mass was incorrect in the ReadForceFieldDefinitionsMixingRules function. This has been corrected for both when GeneralMixingRule=LORENTZ_BERTHELOT, and when GeneralMixingRule=JORGENSEN.

Note: When using both the force_field_mixing_rules.def and force_field.def files to define the force field parameters, you do not encounter this bug as in the ReadForceFieldDefinitions function, the formula to compute the reduced mass is correct.